Added a test for genTxAndNewEpoch. Added Utils
